🐛 Detect missing core `php` formula (#66)
In rare cases, the version corresponding to the `php` formula might not
be installed, but another alias is linked correctly. That means that the
PHP binary is found, but the core formula is not. PHP Monitor will
incorrectly report that it exists, which means the user can break their
PHP installation. Oops.

This quick fix handles that situation:

- Checks if the PHP binary for the `php` aliased version exists
  (located in $optdir/php/bin/php)
- Disables the force load option if the aliased version is missing
  (including a tweaked label if the version is missing)