✨ Query latest version of Valet via Packagist
This makes it possible for PHP Monitor to request what the latest
version of Valet is. This isn't wired up to the UI currently, so
this feature isn't enabled yet.

To enable this feature, I would need to add:
- Conditional check for Valet updates (via setting)
- Decide when to run this check (either every X time and at launch)?

Additionally, PHP Monitor should be aware of its own uptime in order
to make periodic checks if the app hasn't been restarted. A check
should ideally occur every week or so if the app is not restarted.

The relevant app update check should also be adjusted to run in a very
similar way. How frequently the app checks for updates might also be
a setting but I don't want to query too often.