✨ Add randomized UUID for API requests
A description of how this UUID is used (for legitimate purposes, not for
tracking of individual users) is included in the file. Since the API ID
is now also used in API requests, I've bumped the version of the user
agent in the list of default headers sent via the RealWebApi.